requirements.txt
aiohttp>=3.8.0
pywebview>=4.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
open-webui>=0.1.0